% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{id2015}
\alias{id2015}
\title{The 2015 Idaho Snow Load Study Dataset}
\format{A data frame with 651 rows and 9 variables:

 \describe{
 \item{STATION}{The National Weather Service COOP, or Natural Resource
 conservation service Snow Course station identifier.}
 \item{STATION_NAME}{The character name of the station.}
 \item{STATE}{Two character state abbreviation of the
 measurement location.}
 \item{LONGITUDE}{Longitude coordinate position.}
 \item{LATITUDE}{Latitude coordinate position.}
 \item{ELEVATION}{Elevation of the measurement location (meters).}
 \item{YRS}{The number of years of recorded load maximums. This is
 effectively the sample size of the 50 year estimate.}
 \item{yr50}{The estimated design (i.e. 50 year) ground snow load (kpa).}
 \item{HUC}{The United States Geological Survey 8 digit
 hydrologic unit code.}
 }}
\usage{
id2015
}
\description{
50 year ground snow load estimates at 651 measurement locations
in and around the state of Idaho. Used as part of the
2015 Idaho snow load study.
}
\details{
These data consist of 394 (246 SC/SNOTEL, 148 COOP) Idaho stations
with an additional 257 (222 SC/SNOTEL, 35 COOP) located near the Idaho
border with the most recent measurements being taken in 2014. log-Pearson
type III distribution parameter estimates were determined using the sample
mean, standard deviation skew of annual of yearly maximums at each station
location (i.e. method of moments).
}
\references{
\insertRef{Hatailah2015}{snowload}
}
\keyword{datasets}
