% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherMethods.R
\name{idw_snow}
\alias{idw_snow}
\title{Inverse Distance Weighting, the Idaho way}
\usage{
idw_snow(variables = c("RESPONSE", "ELEVATION"), locations, newdata,
  tlayer = 1220, power = c(2, 6), NGSL = TRUE,
  bound_output = FALSE, ...)
}
\arguments{
\item{locations}{The measurement locations to be used for interpolation.
An object of class SpatialPointsDataFrame}

\item{newdata}{The prediction locations. An objet of class
SpatialPointsDataFrame, SpatialGridDataFrame, or SpatialPixelsDataFrame.}

\item{tlayer}{The elevation (in meters) on which to split the data. The
Idaho snow load report splits the state into two layers using
4000 ft (1219.2 meters) as a cutoff.}

\item{power}{A vector of length 2 indicating the exponent in the inverse
distance weighting for the lower and upper elevations respectively.}

\item{NGSL}{If true, used normalized ground snow loads, which are simply
snow loads divided by elevation. This is the way elevation is accounted
for in the 2015 Idaho study.}

\item{bound_output}{If true, the final predictions are capped at the range
of observed data.}

\item{formula}{A formula that is passed to the idw function in the
sp package.}
}
\value{
An appended version of the newdata class containing a column named
  "idw_snow".
}
\description{
A function that implements inverse distance weighting using
normalized ground snow loads.
}
