% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherMethods.R
\name{lm_snow}
\alias{lm_snow}
\title{Linear regression for ground snow load prediction.}
\usage{
lm_snow(formula = log(RESPONSE) ~ ELEVATION, locations, newdata,
  bound_elevation = FALSE, bound_output = FALSE)
}
\arguments{
\item{formula}{A forumla object passed directly to the linear model.}

\item{locations}{An object of class SpatialPointsDataFrame. Specifies
the measurement locations.}

\item{newdata}{A spatial data object. Specifies the prediction
locations.}

\item{bound_elevation}{If TRUE, the elevations of the newdata locations
are restricted to the range of observed data in the locations object.}

\item{bound_ouput}{If TRUE, the final ouputs are restricted to the observed
outputs at the measurement locations.}
}
\value{
A numeric vector of predictions with length equal to the number
  of rows in newdata.
}
\description{
This function is nothing more than a wrapper to the standard linear
models with inputs designed to look and feel similar to other
functions in the snowload package.
}
