% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism.R
\name{prism_tune}
\alias{prism_tune}
\title{Function to tune PRISM parameters using cross validation.}
\usage{
prism_tune(formula = log(RESPONSE) ~ ELEVATION, locations,
  distImp = 0.8, minRad = 10, wdistance = 2, welevRange = list(lwr
  = 200, upr = 2500), welevation = c("ELEVATION", 1), wbasin = c(NA,
  3), weco = c(NA, 3), bound = FALSE, tune_folds = 10,
  tune_score = "MAE")
}
\description{
This function conducts a grid search over the given list of
parameters and selects parameters that minimize the cross validated
error according to the provided scoring method.
}
