% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkriging.R
\name{rkriging}
\alias{rkriging}
\title{Regression-kriging functions for snow load estimation.}
\usage{
rkriging(formula = log(RESPONSE) ~ ELEVATION, locations, newdata, model,
  bound_elevation = FALSE, bound_output = FALSE, sklm = FALSE, ...)
}
\arguments{
\item{formula}{The formula used to account for elevation in the predictions.
Typically of the form given as the default option.}

\item{locations}{The measurement locations (used to train the model).
Of class SpatialPointsDataFrame from the sp package.}

\item{newdata}{The prediction locations. Must be one of the accepted spatial
classes from the sp package.}

\item{model}{A variogram model created with the vgm function
from the gstat package.}

\item{bound_elevation}{If true, restrict the trend predictions using the
range of measurement location elevations.}

\item{bound_output}{If true, restrict the final predictions to the range
of observed data at measurement locations.}

\item{sklm}{If true, use simple kriging with varying local means. If false,
use "universal kriging". The only difference is that sklm assumes no
correlation between locations when fitting the linear model for
elevation.}
}
\value{
An appended version of newdata which includes a column
  named "rkriging" containing the predictions. Note that if
  bound_elevation-TRUE then the elevations in this object might be
  different than the original object.
}
\description{
The function combines simple kriging with varying local means (SKLM) and
universal kriging (UK) (more precisely, kriging with an external drift).
The only difference between the two regression-kriging estimators is that
UK leverages spatial correlations in the regression fitting.
}
