% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{snowload-package}
\alias{snowload-package}
\alias{_PACKAGE}
\title{Snowload prediction methods in R.}
\description{
This package is a compilation of works completed as part of the 
    2018 Ground Snow Load Study. As our methodologies continue to improve, 
    this package will serve as the repository for the functions used to 
    collect, clean, and process data, as well as make predictions between
    measurement locations.
}
\author{
Brennan Bean \email{brennan.bean.20@gmail.com}
}
