% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{snowloads}
\alias{snowloads}
\title{The National Snow Load Dataset}
\format{A data frame with 11,444 rows and 14 variables:

 \describe{
 \item{STATION}{The numeric station identifier. Actual station names are
 not provided as each location may represent more than one station.}
 \item{STATION_NAME}{The character name of the first station
 (in alphabetical order) in the cluster.}
 \item{STATE}{Two character state abbreviation of the
 measurement location. Includes Canadian provinces.}
 \item{LONGITUDE}{Longitude coordinate position.}
 \item{LATITUDE}{Latitude coordinate position.}
 \item{ELEVATION}{Elevation of the measurement location (meters).}
 \item{REGIONIII}{U.S. Environmental Protection Agency level III
 numeric ecoregion.}
 \item{climate}{Climate class based on Sturm et al. (1995).}
 \item{HUC}{The United States Geological Survey 8 digit
 hydrologic unit code.}
 \item{numST}{The number of stations comprising the measurement location.}
 \item{numMax}{The number of annual maximums that
 were retained from the period of record (i.e. the sample size).}
 \item{medMax}{The median of the annual maximum snow loads from the
 historical record.}
 \item{mMax}{The maximum snow load observation from the
 period of record (kpa).}
 \item{snowload}{The estimated design (i.e. 50 year) ground snow load (kpa).}
 }}
\usage{
snowloads
}
\description{
50 year ground snow load estimates at 11,444 measurement locations
in North America.
}
\details{
This dataset was created by scraping data from the global
historical climatological network. Values were subject to automatic
outlier detection and removal. 50 year estimates were obtained by
fitting 5 distributions (log-normal, gamma, extreme value distributions
types I, II, and III) and selecting the median 98th percentile
from the 5 distributions. Climate types were determined using
eco-region maps from the Environmental Protection Agency (EPA)
and climate classes developed by Sturm et al. (1995).

Note that each measurement location may represent more than
one station. When two stations had overlapping period of
records, only the maximum snow load estimate was retained.
Preference was also always given to direct measurements
of snow load.
}
\references{
\insertRef{eco2018}{snowload}
\insertRef{Sturm1995}{snowload}
}
\keyword{datasets}
