% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSpatial.R
\name{toSpatial}
\alias{toSpatial}
\title{Convert snowload datasets to spatial data frames}
\usage{
toSpatial(tdata)
}
\arguments{
\item{tdata}{A string naming one of the datasets internal to the snowload
package.}
}
\value{
A spatial version with the same coordinate extent as the
  utdem SpatialPixelsDataFrame. (Also internal to the snowload package.)
}
\description{
This function is a simple wrapper to sp package functions to create spatial
  data frame objects with a common projection. The function assumes that
  the variable name conventions of the datasets internal to the snowload
  package. It is not intended to be used on datasets outside of the
  snowload package.
}
