% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ut1992}
\alias{ut1992}
\title{The 1992 Utah Snow Load Study Dataset}
\format{A data frame with 413 rows and 10 variables:

 \describe{
 \item{STATION}{The National Weather Service COOP, or Natural Resource
 conservation service Snow Course station identifier.}
 \item{STATION_NAME}{The character name of the station.}
 \item{LONGITUDE}{Longitude coordinate position.}
 \item{LATITUDE}{Latitude coordinate position.}
 \item{ELEVATION}{Elevation of the measurement location (meters).}
 \item{approx}{Indicates whether or not the station location was
 approximated using Google Earth (1-yes, 0-no).}
 \item{maxobs}{The maximum snow load observation from the
 period of record (kpa).}
 \item{yr50}{The estimated design (i.e. 50 year) ground snow load (kpa).}
 \item{HUC}{The United States Geological Survey 8 digit
 hydrologic unit code.}
 }}
\usage{
ut1992
}
\description{
50 year snow load estimates at 413 measurement locations
(210 Snow Course, 203 COOP) in Utah. Used to determine Utah's
1992 ground snow load requirements.
}
\details{
The data include
50 year estimates were determined using a log-Pearson III
applied to annual maximum snow loads. Where required, snow loads were
estimated from snow depths using the rocky mountain conversion density.

Station locations were not provided in the original data. Locations were
inferred from NWS station databases, personal correspondence with the
Utah Snow Survey Office in Salt Lake City, and Google Earth.
}
\references{
\insertRef{SEAU1992}{snowload}
}
\keyword{datasets}
