% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ut2017}
\alias{ut2017}
\title{The 2017 Utah Snow Load Study Dataset}
\format{A data frame with 415 rows and 10 variables:

 \describe{
 \item{STATION}{The National Weather Service COOP, or Natural Resource
 conservation service SNOTEL station identifier.}
 \item{STATION_NAME}{The character name of the station.}
 \item{STATE}{Two character state abbreviation of the
 measurement location.}
 \item{LONGITUDE}{Longitude coordinate position.}
 \item{LATITUDE}{Latitude coordinate position.}
 \item{ELEVATION}{Elevation of the measurement location (meters).}
 \item{YRS}{The number of years of recorded load maximums. This is
 effectively the sample size of the 50 year estimate.}
 \item{maxobs}{The maximum snow load observation from the
 period of record (kpa).}
 \item{yr50}{The estimated design (i.e. 50 year) ground snow load (kpa).}
 \item{HUC}{The United States Geological Survey 8 digit
 hydrologic unit code.}
 }}
\usage{
ut2017
}
\description{
50 year ground snow load estimates at 415 measurement locations in
locations in and around the state of Utah. Used as part of the
2018 Utah snow load study.
}
\details{
This dataset contains 279 (192 COOP, 87 SNOTEL) Utah stations
with an additional 136 stations (103 COOP, 33 SNOTEL), all located within
100 km of the Utah border. 50 year estimates were obtained by fitting
a log-normal distribution to the annual maximum snow loads (by water
year, not calendar year) via maximum likelihood estimation and
extracting the 98th percentile.
}
\references{
\insertRef{Bean2018-report}{snowload}
}
\keyword{datasets}
