% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{utdem}
\alias{utdem}
\title{A 1km by 1km digital elevation model for Utah.}
\format{A SpatialPixelsDataFrame with the following variables.

 \describe{
 \item{HUC}{The United States Geological Survey 8 digit
 hydrologic unit code.}
 \item{ELEVATION}{The elevation of the location (meters).}
 }}
\usage{
utdem
}
\description{
These elevation values were obtained from the U.S.
Geological Survey National Map. They were aggregated from
their original 30m by 30m resolution to a 1km by 1km
resultion using \code{\link[raster]{aggregate}}.
}
\keyword{datasets}
