% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherMethods.R
\name{tri_snow}
\alias{tri_snow}
\title{Linear Triangulation Interpolation}
\usage{
tri_snow(variables = c("RESPONSE", "ELEVATION"), locations, newdata,
  density = c(500, 500), NGSL = TRUE, bound_output = FALSE, ...)
}
\arguments{
\item{variables}{A vector of length two specifying the column names
of the response and elevation variables respectively.}

\item{locations}{The measurement locations to be used for interpolation.
An object of class SpatialPointsDataFrame}

\item{newdata}{The prediction locations. An objet of class
SpatialPointsDataFrame, SpatialGridDataFrame, or SpatialPixelsDataFrame.}

\item{density}{The density of the grid created by the interpolation.
The first argument specifies the x density and the second the y density.}

\item{NGSL}{If true, normalized ground snow loads (which are snow loads
divided by elevation) are used in prediction.}

\item{bound_ouput}{If TRUE, the final ouputs are restricted to the observed
outputs at the measurement locations.}
}
\value{
An appended version of newdata that includes a column named
  "tri_snow" with the predictions.
}
\description{
A function that implements a linear triangulation interpolation.
The function is a wrapper to appropriate functions in the
akima package.
}
